quick-fix - Adjusted suggested office hours
I realized that Tuesday is basically the only day I have a hard commitment to something after 5, so I recommend setting up appointments for Monday, Wednesday, Thursday, or Friday. Of course, if Tuesday's the only time that works for you I'll find a way to make it work for me too.